# examples/openai_api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FunASR OpenAI-Compatible API",
    "version": "1.0.0",
    "description": "Static OpenAPI description for the FunASR OpenAI-compatible speech transcription server. The running FastAPI service also exposes Swagger UI at /docs and an OpenAPI document at /openapi.json."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    },
    {
      "url": "http://funasr-api:8000",
      "description": "Example Docker Compose or internal service name"
    }
  ],
...
  "paths": {
    "/v1/audio/transcriptions": {
      "post": {
        "tags": [
          "Transcription"
        ],
        "summary": "Transcribe an audio file",
        "description": "Upload an audio file with multipart/form-data and receive an OpenAI-compatible transcript response. Set response_format=verbose_json to include segments and timing metadata.",
        "operationId": "createTranscription",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/TranscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transcription result",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TranscriptionTextResponse"
                    },
                    {
                      "$ref": "#/components/schemas/VerboseTranscriptionResponse"
                    }
                  ]
                },
                "examples": {
                  "json": {
                    "summary": "Default text response",
                    "value": {
                      "text": "Welcome to FunASR."
                    }
                  },
                  "verbose_json": {
                    "summary": "Verbose response",
                    "value": {
                      "text": "Welcome to FunASR.",
                      "segments": [
                        {
                          "start": 0.0,
                          "end": 1.6,
                          "text": "Welcome to FunASR.",
                          "speaker": null
                        }
                      ],
                      "language": "auto",
                      "duration": 0.843,
                      "model": "sensevoice"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown model alias",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid multipart request"
          },
          "500": {
            "description": "Transcription runtime error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TranscriptionRequest": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Audio file such as wav, mp3, flac, m4a, ogg, or webm."
          },
          "model": {
            "type": "string",
            "enum": [
              "sensevoice",
              "paraformer",
              "paraformer-en",
              "fun-asr-nano",
              "moss-transcribe-diarize"
            ],
            "default": "sensevoice",
            "description": "FunASR model alias."
          },
          "language": {
            "type": "string",
            "nullable": true,
            "description": "Optional language hint passed to FunASR."
          },
          "response_format": {
            "type": "string",
            "enum": [
              "json",
              "verbose_json"
            ],
            "default": "json",
            "description": "Use verbose_json to include segments and timing metadata."
          }
        }
      },
      "TranscriptionTextResponse": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string",
            "description": "Cleaned transcript text."
          }
        }
      },
      "VerboseTranscriptionResponse": {
        "type": "object",
        "required": [
          "text",
          "segments",
          "language",
          "duration",
          "model"
        ],
        "properties": {
          "text": {
            "type": "string"
          },
          "segments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptionSegment"
            }
          },
          "language": {
            "type": "string",
            "example": "auto"
          },
          "duration": {
            "type": "number",
            "format": "float",
            "description": "Server-side transcription latency in seconds."
          },
          "model": {
            "type": "string",
            "example": "sensevoice"
          }
        }
      },
      "TranscriptionSegment": {
        "type": "object",
        "required": [
          "start",
          "end",
          "text"
        ],
        "properties": {
          "start": {
            "type": "number",
            "format": "float",
            "description": "Segment start time in seconds."
          },
          "end": {
            "type": "number",
            "format": "float",
            "description": "Segment end time in seconds."
          },
          "text": {
            "type": "string"
          },
          "speaker": {
            "type": "string",
            "nullable": true,
            "description": "Speaker label when available."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      }
    }
  }
}